requirements.txt
gradio==6.11.0
fastapi==0.135.3
uvicorn==0.44.0
python-dotenv==1.1.1
pyngrok
crewai==1.14.0
litellm==1.83.0
edge-tts==7.2.8
openai-whisper==20250625
opencv-python==4.13.0.92
pillow==11.3.0
rembg==2.0.69
replicate==1.0.7
openai==2.30.0
google-genai==1.65.0
google-generativeai==0.8.6
groq==0.37.1
langdetect==1.0.9
requests==2.32.5
mysql-connector-python
bcrypt==5.0.0
pydantic==2.11.10
